+++ packages/stylist-core/src/ast/context.rs
@@ -73,9 +73,7 @@ impl<'a> StyleContext<'a> {
 
     /// Calculate the layers that current context differs from the parent context
     fn unique_conditions(&self) -> impl Iterator<Item = &str> {
-        self.conditions()
-            .into_iter()
-            .skip(self.common_conditions().count())
+        self.conditions().skip(self.common_conditions().count())
     }
 
     /// Calculate the layers that parent context differs from current context
@@ -140,7 +138,7 @@ impl<'a> StyleContext<'a> {
             // Use the selector of parent context
             .or_else(|| self.selector.clone())
             // Use class name of scope context
-            .or_else(|| self.class_name.map(|m| format!(".{}", m).into()))
+            .or_else(|| self.class_name.map(|m| format!(".{m}").into()))
             // Use html
             .or_else(|| Some(":root".into()));