+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Changelog
 
+### v0.12.1
+
+- Implemented a workaround that mitigates the linker error from Rust compiler (https://github.com/rust-lang/rust/issues/111888).
+
+## v0.12.0
+
+### Other Changes:
+
+- Added Server-side Rendering Support.
+- Fixed a reference cycle between `Style` and `StyleManager`.
+- Added `StyleManager::new()` to create a style manager with default configuration.
+
 ## v0.11.0
 
 ### Breaking Changes:
+
 - Yew version is bumped to v0.20.
 - Remove `YieldStyle`. This API can be easily reproduced in user code, if need be,
   but often leads to clumsy code in struct components. Use alternative API and prefer
@@ -17,6 +30,7 @@
   runtime.
 
 ### Other Changes:
+
 - The `Style::new_*` API is more open for accepted types of the `Css` parameter.
 - The name of styled components now defaults to the name of the function, like in
   `function_component`.
@@ -24,15 +38,18 @@
 ## v0.10.1
 
 ### Other Changes:
+
 - Added an impl of `IntoPropValue<Classes>` for `Style` and `StyleSource` when
   the `yew_integration` feature is active.
 
 ## v0.10.0
 
 ### Breaking Changes:
+
 - Yew version is bumped to 0.19.
 
 ### Other Changes:
+
 - Added an API to style Yew Function Component.
 - `random` features is now provided with `fastrand`.
 - Added Yew hooks for Media Query.
@@ -41,16 +58,19 @@
 ## v0.9.2
 
 ### Other Changes:
+
 - Fixed a misconfiguration causing documentation fails to build on `docs.rs`.
 
 ## v0.9.1
 
 ### Other Changes:
+
 - Removed an unused import.
 
 ## v0.9
 
 ### Breaking Changes:
+
 - `Style` and `GlobalStyle` no longer implements `FromStr`.
 - `Style` and `GlobalStyle` now takes any type that implements
   `Into<StyleSource>` as a source for a stylesheet.
@@ -60,6 +80,7 @@
   `@media`.
 
 ### Other Changes:
+
 - Added a Procedural Macro API that parses the Stylesheet at the compile
   time.
 - Parser will now check stylesheets more strictly.
@@ -75,9 +96,10 @@
 ## v0.8
 
 ### Breaking Changes:
+
 - `Style::new()` and `Style::create()` now takes a new trait `IntoSheet` for
   Stylesheet which is implemented by default for both
-`stylist::ast::Sheet` and everything that implements `AsRef<str>`.
+  `stylist::ast::Sheet` and everything that implements `AsRef<str>`.
 - Feature `yew` has been renamed back to `yew_integration`.
 - Selectors list now gets a class name added for each selector.
 - `Style` is now `!Send` and `!Sync`.
@@ -85,6 +107,7 @@
   and styled-components.
 
 ### Other Changes:
+
 - Added a `GlobalStyle` struct to register global styles.
 - Added a `<Global />` Component for global styling for yew applications.
 - Supported `@supports` CSS at-rule.
@@ -100,6 +123,7 @@
 ## v0.7
 
 ### Breaking Changes:
+
 - `Style::new()` now takes an `Into<Cow<'static, str>>` instead of
   `Into<String>` and returns `stylist::Error` instead of `String` when
   encountering an error.
@@ -109,6 +133,7 @@
 - `Style` no longer implements `ToString`.
 
 ### Other Changes:
+
 - Added a new API `YieldStyle`.
 - Added theming examples.
 - Styles are now cached by default.
@@ -118,15 +143,16 @@
 - Removed Unnecessary Clones.
 - Optimised for Performance.
 
-
 ## v0.6
 
 ### Breaking Changes:
+
 - `style.get_class_name()` no longer consumes the style and returns a `&str`
   instead of an owned string.
 - Seed Integration is Removed.
 
 ### Other Changes:
+
 - Added `Style::new` which does not require a component name.
 - Aesthetically pleasing Class Name.
 - Replaced `lazy_static` with `once_cell`.